Fix uv readthedocs (#33)
* point to correct conf.py location

* earlier, simpler version from readthedocs doc

* relax readthedocs fail_on_warning

We have lots of issues in the docs at the moment, including example notebooks that are broken. For now, I just want to get the docs building again, so I'm going to disable the fail_on_warning option. I'll come back to this later.

* changed example ipynb name to avoid conflict

* fix link to directivity example

Also upgraded directivity_example.ipynb to jupyter v4

* add readthedocs badge