FIX: avoid numpy 2.0 deprecation warning
From v2, `__array_wrap__` needs `return_scalar` argument. Implemented with a check for numpy version for backwards compatibility to numpy v1.
Looks like this will also be deprecated in the future, so this is a bit of a hack, but it will all change if we change to `xarray` anyway.